debian: put mod_ilbc in non-free
The iLBC library in our tree is non-free software.  Google has
purchased Global IP Sound and released a newer version of this library
as part of WebRTC as free software, but we need to upgrade to take
advantage of this more favorable licensing.

Until we do that, we also need to remove libs/ilbc from the Debian
source packaging.