use the typedef for an enum rather than unsigned
The callback here receives both zrtp_protocol_event_t and
zrtp_security_event_t events.  The latter is numbered contiguously to
the former.  Still, it's a bit confusing to have to pick one as a
type, but apparently this fixes a warning on Visual Studio on Windows.

Thanks to Jeff Lenk.