Fix handling of the ZRTP cache path and file name
libzrtp strings are weird, and we were previously misusing the
function zrtp_zstrcpyc.  We can't use ZSTR_GV because it does insane
things and causes an array-bounds warning on gcc 4.7.  So we have to
take matters into our own hands and setup the string correctly and
copy data into it.

Because we were doing it wrong, people would get weird pseudo-random
single-character names for the zrtp cache file, and the file would end
up in the wrong place.  Now that this is fixed, users will need to
locate and move their zrtp cache file to their db_dir and name it
"zrtp.dat" if they wish to keep their current ZRTP cache.

FS-4344 --resolve